Update Login Screen Functionality
Update login screen so that it checks whether the user
that is logging in is a customer or an admin. If the user is
an admin, send the user to MainActivity.class; if the user is
a customer, send the user to MainActivityUser.class (once it
is created, code temporarily sends the user back to splash
screen).