Implement Account Summary and Close Account
Connect MainActivity to DisplayUserInfoActivity. DisplayUserInfoActivity still
displays upon registration of an account, but it will also check whether it was
accessed after account registration or when the user clicks account summary.

Clicking Close Account will currently delete both the logged in User as well as
all the Accounts associated with his or her username.